# src/models/requests/update_todo_request.go
package requests

// UpdateTodoRequest is the payload for PUT /todos/:id.
type UpdateTodoRequest struct {
	Title       string `json:"title"`
	Description string `json:"description"`
	Completed   bool   `json:"completed"`
}